requirements.txt
langchain
langchain-google-genai
PyGithub
python-dotenv
fastapi
uvicorn

